taipy_designer_demos/6_genesis/requirements.txt
geopandas==0.14.4
requests==2.32.3
plotly==5.22.0